+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/SplashTheme"
         tools:targetApi="31">
         <activity
-            android:name=".BoardActivity"
+            android:name=".community.ui.BoardActivity"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
