+++ app/src/main/java/com/example/hrr_android/community/ui/fragment/CommunityTabContentFragment.kt
@@ -1,4 +1,4 @@
-package com.example.hrr_android
+package com.example.hrr_android.community.ui.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,7 +7,8 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.example.hrr_android.databinding.FragmentCommunityContentBinding
+import com.example.hrr_android.community.model.Post
+import com.example.hrr_android.community.adapter.PostRVAdapter
 import com.example.hrr_android.databinding.FragmentCommunityTabContentBinding
