+++ app/src/main/res/layout/fragment_community_content.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CommunityContentFragment">
+    tools:context=".community.ui.fragment.CommunityContentFragment">
 
     <include
         android:id="@+id/layout_board_hot_post"
