+++ app/src/main/res/layout/fragment_community_tab_content.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CommunityTabContentFragment">
+    tools:context=".community.ui.fragment.CommunityTabContentFragment">
 
     <!-- 선 (구분선) -->
     <View
